+++ Cargo.toml
@@ -1,7 +1,15 @@
 [package]
-name = "cook-import"
+name = "cooklang-import"
 version = "0.1.0"
 edition = "2021"
+
+[lib]
+name = "cooklang_import"
+path = "src/lib.rs"
+
+[[bin]]
+name = "cooklang-import"
+path = "src/main.rs"
 
 [dependencies]
 env_logger = "0.11.5"
